+++ src/nationalarchives/components/cookie-banner/cookie-banner.stories.js
@@ -13,6 +13,10 @@ const argTypes = {
   cookiesDomain: { control: "text" },
   cookiesPath: { control: "text" },
   allowInsecure: { control: "boolean" },
+  style: {
+    control: "inline-radio",
+    options: ["none", "contrast", "accent", "tint"],
+  },
   classes: { control: "text" },
   attributes: { control: "object" },
 };
@@ -37,6 +41,7 @@ const Template = ({
   cookiesDomain,
   cookiesPath,
   allowInsecure,
+  style,
   classes,
   attributes,
 }) =>
@@ -50,6 +55,7 @@ const Template = ({
       cookiesDomain,
       cookiesPath,
       allowInsecure,
+      style,
       classes,
       attributes,
     },
@@ -66,6 +72,7 @@ Accept.args = {
   serviceName: "My service",
   cookiesUrl: "#",
   allowInsecure: true,
+  style: "contrast",
   classes: "tna-cookie-banner--demo",
 };
 Accept.play = async ({ canvasElement }) => {
@@ -106,6 +113,7 @@ export const Reject = Template.bind({});
 Reject.args = {
   serviceName: "My service",
   cookiesUrl: "#",
+  style: "contrast",
   classes: "tna-cookie-banner--demo",
 };
 Reject.play = async ({ canvasElement }) => {
@@ -143,6 +151,7 @@ CustomPolicies.args = {
   serviceName: "My service",
   cookiesUrl: "#",
   policies: "custom",
+  style: "contrast",
   classes: "tna-cookie-banner--demo",
 };
 CustomPolicies.parameters = {
@@ -178,6 +187,7 @@ Existing.args = {
   serviceName: "My service",
   cookiesUrl: "#",
   allowInsecure: true,
+  style: "contrast",
   classes: "tna-cookie-banner--demo",
 };
 Existing.decorators = [